refactor: dto 분리
- dto를 comment와 post로 분리
