+++ src/main/java/io/sparta/board/application/dto/request/comment/CommentCreateRequestDto.java
@@ -1,18 +1,20 @@
-package io.sparta.board.application.dto.response;
+package io.sparta.board.application.dto.request.comment;
 
+import io.sparta.board.domain.entity.Comment;
 import io.sparta.board.domain.entity.Post;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.UUID;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class CommentCreateResponseDto {
-    private UUID id;
+public class CommentCreateRequestDto {
     private String content;
+
+    public Comment toEntity(Post post) {
+        return Comment.create(content, post);
+    }
 }
