+++ src/main/java/io/sparta/board/application/dto/request/post/PostUpdateRequestDto.java
@@ -1,19 +1,16 @@
-package io.sparta.board.application.dto.response;
+package io.sparta.board.application.dto.request.post;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-import java.util.UUID;
-
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class CommentResponseDto {
-    private UUID id;
+public class PostUpdateRequestDto {
+    private String title;
     private String content;
-    private LocalDateTime createdAt;
+
 }
