+++ src/main/java/io/sparta/board/application/dto/response/comment/CommentResponseDto.java
@@ -1,0 +1,19 @@
+package io.sparta.board.application.dto.response.comment;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class CommentResponseDto {
+    private UUID id;
+    private String content;
+    private LocalDateTime createdAt;
+}
